Extract classes for tests
